requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
anyio==3.6.2
astroid==2.13.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
databases==0.7.0
dill==0.3.6
distlib==0.3.6
fastapi==0.89.1
fastapi-cache2==0.2.0
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.15
idna==3.4
iniconfig==2.0.0
isort==5.11.4
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
pathspec==0.10.3
pendulum==2.1.2
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.1
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
pylint==2.15.10
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytzdata==2020.1
PyYAML==6.0
redis==4.4.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomlkit==0.11.6
types-pyOpenSSL==23.0.0.2
types-redis==4.4.0.3
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.17.1
wrapt==1.14.1
yarl==1.8.2
